# Faster by concurrently parsing the OS, UA & Device