* Gain of 33%
* `uap-go` is often used on server side to parse agents from incoming http requests i.e. high volume. Making it a bit faster while preserving Goroutine safety is therefore useful.
```
benchmark             old ns/op     new ns/op     delta
BenchmarkParser-8     27243975      18124922      -33.47%
```